todakhanip-api/requirements.txt
pymongo==4.6.1
flask==3.0.0
openai==1.12.0
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0 